requirements.txt
django == 5.0.7
djangorestframework == 3.15.2
requests == 2.32.3
#djangorestframework-simplejwt == 5.3.1
mssql-django == 1.5